requirements.txt
django
python-dotenv #for secure data
psycopg2 #server side needed
pillow #for images
django-bootstrap-v5 #nice looking CSS
django-crispy-forms #forms for login/register, automate the work
crispy-bootstrap5 #forms for login/register, automate the work
pyyaml # working with django-user-agents
ua-parser # working with django-user-agents
user-agents # working with django-user-agents
django-user-agents # working with django-user-agents
django-recaptcha #safety for login
django-registration #two-phase registration
django-ipware #to receive client's IP - make sure no spamming bots
django-image-cropping #crop added images
easy-thumbnails #crop functions work with django-image-cropping